image texture

+++ ray-tracing-core/include/math/functions.h
@@ -21,7 +21,7 @@ namespace ray_tracing_core::math
     {
         auto r0 = (1.0 - ref_idx) / (1.0 + ref_idx);
         r0 = r0 * r0;
-        return r0 + (1.0 - r0) * std::powf(1.0 - cosine, 5.0);
+        return r0 + (1.0 - r0) * std::pow(1.0 - cosine, 5.0);
     }
 }
